Use direct archetype match when relevant
We used to check against every filter conjunction for each query. This
is because tick filters (Added, Changed) need to check for tick only on
a subset of archetypes.

Now, instead of checking every filter conjunction, we only check:

If, and only if, there is a tick filter in a conjunction, check
That the filter is satisfied (we can assume the fetch already is),
then check the tick.

Now archetype checks are trivial (usually a bitset `contains`) op. Same
for query iterations.